refactor(dashboard): align auth forms with reinhardt 0.2
Share auth registration and session-cookie helpers across REST and server_fn flows.

Keep auth pages on static form! definitions and redirect successful registration to login.
