+++ dashboard/src/apps/auth/client/pages/login.rs
@@ -1,8 +1,8 @@
 //! Login page with username and password form.
 //!
-//! Uses `form!` macro for declarative form rendering with automatic
-//! server function integration. On successful login, the auth state is
-//! updated via `AuthState` and the user is redirected to the dashboard.
+//! Uses `form!` for the static field/server_fn definition. Login success is
+//! persisted by the server-side session cookie, then the form redirects to the
+//! dashboard route.
 
 use reinhardt::pages::component::Page;
 use reinhardt::pages::form;
@@ -16,6 +16,7 @@ pub fn login_page() -> Page {
 	let login_form = form! {
 		name: LoginForm,
 		server_fn: login,
+		method: Post,
 		class: "space-y-4",
 		redirect_on_success: "/",
 		fields: {
