+++ dashboard/src/apps/auth/views/register.rs
@@ -3,20 +3,15 @@
 //! Creates a new user with `is_active = false` and sends a verification
 //! email. The user must verify their email before they can log in.
 
-use reinhardt::core::exception::Error as AppError;
 use reinhardt::core::serde::json;
-use reinhardt::db::orm::Model;
 use reinhardt::di::Depends;
 use reinhardt::http::ViewResult;
 use reinhardt::post;
-use reinhardt::{BaseUser, Json, Response, StatusCode};
-use tracing::{error, info};
+use reinhardt::{Json, Response, StatusCode};
 
-use crate::apps::auth::models::User;
 use crate::apps::auth::serializers::RegisterRequest;
 use crate::apps::auth::services::email::EmailService;
-use crate::apps::auth::services::registration::provision_personal_organization;
-use crate::apps::auth::services::token::{TokenPurpose, generate_token};
+use crate::apps::auth::services::registration::register_inactive_user;
 use crate::shared::AuthResponse;
 
 /// Register new user with email verification required.
@@ -28,84 +23,13 @@ pub async fn register(
 	Json(body): Json<RegisterRequest>,
 	#[inject] email_service: Depends<EmailService>,
 ) -> ViewResult<Response> {
-	let settings = crate::config::settings::get_settings();
-	let secret_key = settings.core.secret_key.clone();
-
-	// Create user as inactive — requires email verification to activate
-	let mut user = User::build()
-		.username(body.username.trim().to_string())
-		.email(body.email.trim().to_lowercase())
-		.first_name(String::new())
-		.last_name(String::new())
-		.password_hash(None)
-		.is_active(false)
-		.is_staff(false)
-		.is_superuser(false)
-		.finish();
-	user.set_password(&body.password).map_err(|e| {
-		error!("Password hashing failed during registration: {e}");
-		AppError::Internal("Internal server error".to_string())
-	})?;
-
-	// Attempt to create -- database unique constraint prevents duplicates
-	let created = match User::objects().create(&user).await {
-		Ok(user) => user,
-		Err(e) => {
-			// Normalize error message to lowercase for case-insensitive matching.
-			// The ORM (reinhardt-db) maps unique constraint violations to
-			// `DatabaseError::QueryError(String)` without a structured variant,
-			// so string matching is the only detection mechanism available.
-			let err_lower = e.to_string().to_lowercase();
-			if err_lower.contains("unique") || err_lower.contains("duplicate") {
-				// Distinguish which field caused the conflict by checking
-				// the PostgreSQL constraint name embedded in the error message.
-				let message = if err_lower.contains("auth_user_email_uniq") {
-					"Email already exists"
-				} else {
-					"Username already exists"
-				};
-				return Err(AppError::Conflict(message.to_string()));
-			}
-			error!("Failed to create user in database: {e}");
-			return Err(AppError::Internal("Internal server error".to_string()));
-		}
-	};
-
-	// Provision a Personal Organization for the new user (refs #415, #435).
-	// Slug derivation, retry, and rollback are handled by the shared service.
-	provision_personal_organization(&created).await?;
-
-	// Generate verification token and send email
-	let token = generate_token(
-		TokenPurpose::EmailVerification,
-		&created.id,
-		"",
-		&secret_key,
-	);
-
-	let port = std::env::var("PORT").unwrap_or_else(|_| "8000".to_string());
-	let base_url = std::env::var("REINHARDT_CLOUD_BASE_URL")
-		.unwrap_or_else(|_| format!("http://localhost:{port}"));
-	let verification_url = format!("{base_url}/api/auth/verify-email/{token}/");
-
-	if let Err(e) = email_service
-		.send_verification_email(&created.email, &created.username, &verification_url)
-		.await
-	{
-		error!(
-			"Failed to send verification email to {}: {e}",
-			created.email
-		);
-		// Roll back user creation to avoid stranding an inactive account
-		// with no way to receive the activation link.
-		if let Err(del_err) = User::objects().delete(created.id).await {
-			error!("Failed to roll back user after email failure: {del_err}");
-		}
-		return Err(AppError::Internal(
-			"Registration failed — please try again later".to_string(),
-		));
-	}
-	info!("Verification email sent to {}", created.email);
+	let created = register_inactive_user(
+		&body.username,
+		&body.email,
+		&body.password,
+		email_service.as_ref(),
+	)
+	.await?;
 
 	let resp = AuthResponse {
 		success: true,
